requirements.txt
langgraph>=0.0.48
langchain-google-genai>=1.0.3
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
gradio>=4.20.0
requests>=2.31.0